# openapi-solicitudes-academicas.yaml
openapi: 3.0.3
info:
  title: API — Sistema de Triage y Gestión de Solicitudes Académicas
  description: |
    API REST para el módulo de gestión del ciclo de vida completo de solicitudes académicas
    del Programa de Ingeniería de Sistemas y Computación.

    ## Flujo de estados soportado
    ```
    REGISTRADA → CLASIFICADA → EN_ATENCION → ATENDIDA → CERRADA
                                    ↓               ↓
                                 VENCIDA        IMPUGNADA
                                    ↓
                                 ESCALADA
    ```

    ## Autenticación
    Todos los endpoints requieren un token JWT en el header `Authorization: Bearer <token>`.

    ## Control de concurrencia
    Las operaciones de escritura utilizan **Optimistic Locking** mediante el campo `version`.
    Si se recibe un `409 Conflict`, el cliente debe recargar el recurso y reintentar.

  version: 2.0.0
  contact:
    name: Equipo de Desarrollo — PISC
    email: dev-pisc@universidad.edu.co
  license:
    name: MIT
servers:
  - url: https://api.pisc.universidad.edu.co/v2
    description: Producción
  - url: https://api-staging.pisc.universidad.edu.co/v2
    description: Staging
  - url: http://localhost:8080/
    description: Desarrollo local
...
paths:
...
  /tipos-solicitud:
    get:
      tags: [Tipos de Solicitud]
      summary: Listar tipos de solicitud activos
      operationId: listarTiposSolicitud
      security:
        - BearerAuth: []
      parameters:
        - name: activo
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Lista de tipos de solicitud
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TipoSolicitudResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Tipos de Solicitud]
      summary: Crear nuevo tipo de solicitud
      description: Requiere rol `DIRECTOR` o `ADMINISTRATIVO`.
      operationId: crearTipoSolicitud
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearTipoSolicitudRequest'
      responses:
        '201':
          description: Tipo creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TipoSolicitudResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /tipos-solicitud/{id}:
    put:
      tags: [Tipos de Solicitud]
      summary: Actualizar tipo de solicitud
      operationId: actualizarTipoSolicitud
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearTipoSolicitudRequest'
      responses:
        '200':
          description: Tipo actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TipoSolicitudResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Tipos de Solicitud]
      summary: Desactivar tipo de solicitud (soft delete)
      description: |
        No elimina físicamente el registro. Marca el tipo como `activo: false`.
        No se puede desactivar un tipo con solicitudes activas asociadas.
      operationId: desactivarTipoSolicitud
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Tipo desactivado exitosamente
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  # ─────────────────────────────────────────────
  # USUARIOS
  # ─────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en el endpoint de autenticación institucional.

  # ─────────────────────────────────────────────
  # PARAMETERS
  # ─────────────────────────────────────────────
...
  schemas:
    EstadoSolicitud:
      type: string
      description: Estado del ciclo de vida de la solicitud
      enum:
        - REGISTRADA
        - CLASIFICADA
        - EN_ATENCION
        - ATENDIDA
        - CERRADA
        - RECHAZADA
        - CANCELADA
        - VENCIDA
        - ESCALADA
        - IMPUGNADA
      example: CLASIFICADA
...
    CrearTipoSolicitudRequest:
      type: object
      required:
        - nombre
        - descripcion
        - tiempoAtencionDias
      properties:
        nombre:
          type: string
          minLength: 3
          maxLength: 100
          example: "Homologación de asignaturas"
        descripcion:
          type: string
          maxLength: 500
        tiempoAtencionDias:
          type: integer
          minimum: 1
          maximum: 90
          description: SLA en días hábiles para este tipo
          example: 15
...
    TipoSolicitudResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
          example: "Homologación de asignaturas"
        descripcion:
          type: string
        tiempoAtencionDias:
          type: integer
          example: 15
        activo:
          type: boolean
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2024-03-15T10:23:41Z"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        code:
          type: string
          description: Código interno de error para soporte
          example: "SOL-001"
        message:
          type: string
          example: "La descripción de la solicitud es obligatoria y debe tener mínimo 30 caracteres."
        path:
          type: string
          example: "/v2/solicitudes"
        details:
          type: array
          description: Detalle de errores de validación por campo
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      properties:
        field:
          type: string
          example: "descripcion"
        rejectedValue:
          type: string
          nullable: true
          example: "Necesito cupo"
        message:
          type: string
          example: "La descripción debe tener entre 30 y 2000 caracteres"
    ConflictResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 409
        error:
          type: string
          example: "Conflict"
        code:
          type: string
          description: |
            Códigos posibles:
            - TRANSICION_INVALIDA: el estado actual no permite esta operación
            - OPTIMISTIC_LOCK: versión desactualizada, recargar el recurso
            - SOLICITUD_INMUTABLE: la solicitud está en estado final
            - RECURSO_EN_USO: el tipo tiene solicitudes activas
        message:
          type: string
          example: "No es posible clasificar una solicitud en estado EN_ATENCION."
        estadoActual:
          $ref: '#/components/schemas/EstadoSolicitud'
          nullable: true
        path:
          type: string
...
  responses:
    BadRequest:
      description: Parámetros o body inválidos (validación fallida)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-03-15T10:23:41Z"
            status: 400
            error: "Bad Request"
            code: "VAL-001"
            message: "Errores de validación en la solicitud"
            path: "/v2/solicitudes"
            details:
              - field: "descripcion"
                rejectedValue: "Cupo"
                message: "La descripción debe tener entre 30 y 2000 caracteres"
              - field: "solicitanteId"
                rejectedValue: "abc"
                message: "El identificador debe contener solo dígitos"
    Unauthorized:
      description: Token JWT ausente, inválido o expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-03-15T10:23:41Z"
            status: 401
            error: "Unauthorized"
            code: "AUTH-001"
            message: "Token JWT inválido o expirado."
            path: "/v2/solicitudes"
    Forbidden:
      description: El usuario no tiene el rol requerido para esta operación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-03-15T10:23:41Z"
            status: 403
            error: "Forbidden"
            code: "AUTH-002"
            message: "Se requiere rol COORDINADOR para clasificar solicitudes."
            path: "/v2/solicitudes/4821/clasificar"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-03-15T10:23:41Z"
            status: 404
            error: "Not Found"
            code: "SOL-404"
            message: "No se encontró la solicitud con ID 9999."
            path: "/v2/solicitudes/9999"
    Conflict:
      description: Conflicto de estado (transición inválida, optimistic lock o recurso inmutable)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConflictResponse'
          examples:
            transicionInvalida:
              summary: Transición de estado no permitida
              value:
                timestamp: "2024-03-15T10:23:41Z"
                status: 409
                error: "Conflict"
                code: "TRANSICION_INVALIDA"
                message: "No es posible clasificar una solicitud en estado EN_ATENCION."
                estadoActual: "EN_ATENCION"
                path: "/v2/solicitudes/4821/clasificar"
            optimisticLock:
              summary: Versión desactualizada
              value:
                timestamp: "2024-03-15T10:23:41Z"
                status: 409
                error: "Conflict"
                code: "OPTIMISTIC_LOCK"
                message: "La solicitud fue modificada por otro usuario. Recargue y reintente."
                estadoActual: null
                path: "/v2/solicitudes/4821/clasificar"
            solicitudInmutable:
              summary: Solicitud en estado final
              value:
                timestamp: "2024-03-15T10:23:41Z"
                status: 409
                error: "Conflict"
                code: "SOLICITUD_INMUTABLE"
                message: "La solicitud SOL-2024-04821 está CERRADA y no puede ser modificada."
                estadoActual: "CERRADA"
                path: "/v2/solicitudes/4821/cerrar"
...